+++ js/form.js
@@ -1,6 +1,7 @@
 import { resetEffect } from './effects.js';
 import { resetScale } from './image-scale.js';
 
+const FILE_TYPES = ['jpg', 'jpeg', 'png'];
 const HASHTAG_COUNT_MAX = 5;
 const VALID_SYMBOLS = /^#[a-zа-яё0-9]{1,19}$/i;
 const HASHTAG_ERROR_TEXT = `Хэш-тег должен начинаться со знака #, содержать только кириллицу или латиницу, длину от 1-19 символов не включая знак #. Максимальное количество хэштегов: ${HASHTAG_COUNT_MAX}`;
@@ -8,10 +9,12 @@ const HASHTAG_ERROR_TEXT = `Хэш-тег должен начинаться со
 const form = document.querySelector('.img-upload__form');
 const overlay = document.querySelector('.img-upload__overlay');
 const body = document.querySelector('body');
-const uploadFile = document.querySelector('#upload-file');
 const cancelUpload = document.querySelector('#upload-cancel');
 const hashtagField = document.querySelector('.text__hashtags');
 const commentField = document.querySelector('.text__description');
+const fileChooser = document.querySelector('.img-upload__input');
+const preview = document.querySelector('.img-upload__preview img');
+const effectsPreview = document.querySelectorAll('.effects__preview');
 
 const pristine = new Pristine(form, {
   classTo: 'img-upload__field-wrapper',
@@ -66,10 +69,6 @@ const onCancelButtonClick = () => {
   closeModal();
 };
 
-const onFileInputChange = () => {
-  openModal();
-};
-
 pristine.addValidator(hashtagField, validateTags, HASHTAG_ERROR_TEXT);
 
 pristine.addValidator(commentField, validateDescription, 'Длина комментария больше 140 символов');
@@ -85,8 +84,24 @@ const onFormSubmit = (cb) => {
   });
 };
 
+const loadImagePreview = () => {
+  fileChooser.addEventListener('change', () => {
+    const uploadImage = fileChooser.files[0];
+    const uploadImageName = uploadImage.name.toLowerCase();
+
+    const matches = FILE_TYPES.some((it) => uploadImageName.endsWith(it));
+
+    if (matches) {
+      preview.src = URL.createObjectURL(uploadImage);
+      effectsPreview.forEach((item) => {
+        item.style.backgroundImage = `url(${preview.src})`;
+      });
+      openModal();
+    }
+  });
+};
+
 form.addEventListener('submit', onFormSubmit);
-uploadFile.addEventListener('change', onFileInputChange);
 cancelUpload.addEventListener('click', onCancelButtonClick);
 
-export { onFormSubmit, closeModal };
+export { onFormSubmit, closeModal, loadImagePreview };
